requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
google-generativeai==0.3.2
feedparser==6.0.10
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
APScheduler==3.10.4
python-telegram-bot==20.5
gunicorn==21.2.0